SCALE-SEAM (91) — the R20 CRE deal desk, and a forecast tested instead of executed
Thirteen methods out of `client.ts` (867 -> 746) into a new
`apps/web/src/api/creDeal.ts`, lifted as one contiguous run. A mixin, so every
call site resolves unchanged; `api/surface.test.ts` proves it.

What they answer: should we transact on this income property, and on what terms?
Verify the seller's numbers (normalizeT12 — the tie-out is a gate, not a report;
rentRollScrub — a check without its inputs reports not-run, never a pass;
netEffectiveRent; tieredComps; competitiveSupply), decide (holdSell,
decisionGate), and set the terms (clausePlaybook + reviewContractClauses,
loanCovenants, dealAuthority, plus the two save halves).

THREE SIGNALS, DERIVED INDEPENDENTLY, AND THEY AGREED — more than any slice in
this sequence has had:

1. The source marks them. Ten `CRE-` codes sit on these doc comments, and
   grepping the whole of apps/web/src showed those eleven occurrences were the
   ONLY ones in the web tree, every one of them in client.ts.
2. A 1:1 router match. Every route these thirteen call — eleven distinct paths,
   since the playbook and the authority table each have a read and a write half
   on one path — is served by aec_api/routers/realestate.py and by no other
   router, and its ten `(R20)` docstrings name the same ten codes. Checked by
   pairing each @router. line with the def beneath it.
3. One question runs through all thirteen.

A marker alone would not have been enough: four shared-word traps
(entitlements, view, carbon, lifecycle) are exactly why CRE- counts as one vote
of three and the router match carries the weight.

A FOURTH CORROBORATION TURNED UP AFTER THE SLICE WAS COMMITTED, and is recorded
as a fourth rather than folded into the three so the derivation stays honest
about when each piece arrived. The backend suite partitions the same thirteen:
test_cre_deal_desk (comps/tiered, rent-roll/scrub, t12/normalize),
test_cre_governance (deal-room/authority, decision-gate, loan/covenants,
supply/competitive), test_cre_tier3 (contracts/playbook, contracts/review,
hold-sell) and test_net_effective (rent-roll/net-effective) — all eleven paths,
with no test_cre_* file reaching a route outside the set. Two caveats kept in
the header, because this is the kind of late evidence it is tempting to round
up: test_cre_tier3 also exercises /reports/ic, and test_net_effective sets up
through /rent-roll and /modules/lease. It was noticed only because two of those
names scrolled past in the suite output. Test-file names are a grouping somebody
else authored, which is the kind of source (87) argued to prefer — and it was
not on the list of places this slice thought to look.

THE BOUNDARY THAT DECIDES rent-roll. proforma.ts already holds
GET /projects/{pid}/rent-roll, and two of these sit directly beneath it on
/rent-roll/scrub and /rent-roll/net-effective, so a prefix split would have
taken all three. The backend states the boundary: the plain rent roll carries no
CRE- code and no (R20), and its docstring calls it the operating rent roll "from
the lease module (the hold phase)", while these two POST a body of figures the
counterparty supplied. What are we earning and is their number true are
different questions that happen to share a prefix.

AND (88)'S FORECAST WAS TESTED RATHER THAN EXECUTED, AND IT LOST. (88) left
camReconciliation with a note saying it "goes with rentRollScrub,
netEffectiveRent and normalizeT12 ... when a rent-roll slice takes them". This
is that slice; it took all three and left camReconciliation where it was,
because every signal points the other way — no CRE- code, no (R20), and
/projects/{pid}/cam/reconciliation is served by operations.py. A CAM true-up
bills a COMPLETED OPERATING YEAR to sitting tenants; the three it was predicted
to join test a counterparty's figures before a purchase. A placement forecast
written by one slice is a hypothesis for the next slice to test, not an
instruction to carry out — phrased as a plan it reads as settled and invites the
next reader to execute it without re-deriving anything.

THE BANNER CITED A DEPARTED METHOD FOR THE SECOND TIME. rentRollScrub was in the
STAYING banner's example list until this slice moved it, the same defect (88)
caught with esgSummary. (89) audited those seven names and found none had moved,
which is why they survived to be wrong now: a list that passes one audit is not
thereby safe for the next slice.

Two stale numbers in docs/roadmap.md corrected on the way: "client.ts is 953
lines" (three slices out of date) and a lane-table cell still saying (81)-(87),
which is the third recorded instance of that same cell drifting.

Verified: tsc 0, eslint 0, vitest 203 files / 2059 tests, ruff check ../.. clean,
backend suite 660/660, and test_file_sizes / test_claude_md_gates /
test_roadmap_status / test_ruff_scope / test_declared_imports all exit 0.

Pin 867 -> 746. 88 above the banner, still no map.
